+++ selfdrive/loggerd/SConscript
@@ -1,7 +1,9 @@
 Import('env', 'arch', 'cereal', 'messaging', 'common', 'visionipc', 'gpucommon')
 
-src = ['loggerd.cc', 'logger.cc']
-libs = ['zmq', 'capnp', 'kj', 'z',
+
+logger_lib = env.Library('logger', ["logger.cc"])
+src = ['loggerd.cc']
+libs = [logger_lib, 'zmq', 'capnp', 'kj', 'z',
         'avformat', 'avcodec', 'swscale', 'avutil',
         'yuv', 'bz2', 'OpenCL', common, cereal, messaging, visionipc]
 
@@ -22,3 +24,4 @@ if arch == "Darwin":
   env['FRAMEWORKS'] = ['OpenCL']
 
 env.Program(src, LIBS=libs)
+env.Program('bootlog.cc', LIBS=libs)
